+++ astroquery/cadc/core.py
@@ -457,8 +457,7 @@ class CadcClass(BaseQuery):
                                session=self.cadcdatalink._session)
             for service_def in datalink.bysemantics('#cutout'):
                 access_url = service_def.access_url
-                if isinstance(access_url, bytes):  # ASTROPY_LT_4_1
-                    access_url = access_url.decode('ascii')
+
                 if '/sync' in access_url:
                     service_params = service_def.input_params
                     input_params = {param.name: param.value